Add the ReferenceCount value to ExpUnreferencedSampleFilesTable
